requirements.txt
asgiref==3.5.2
Django==4.0.5
Pillow==9.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
sqlparse==0.4.2
web3==5.29.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0